cookbook/models/google/requirements.txt

python-dotenv
google-genai